renamed `platforms` to `os` to better adhere to standards
